requirements.txt
flask==2.3.3
werkzeug==2.3.7
mediapipe==0.10.21
opencv-python==4.8.1.78
numpy==1.26.4
pyrebase4==4.7.1
python-dotenv==0.19.0
websockets==10.0
pyautogui==0.9.53
requests==2.29.0
python-jose==3.3.0
gunicorn==20.1.0 